third_party/blink/web_tests/external/wpt/mediacapture-streams/MediaDevices-after-discard.https-expected.txt

This is a testharness.js-based test.
[FAIL] enumerateDevices() in a discarded browsing context
  assert_equals: Promise state expected "pending" but got "rejected"
Harness: the test ran to completion.
